Batch 19: Audio/gamelogic/control_bar/terrain build fixes — 32 files
- Audio: rubato SincInterpolationParameters, rodio f32 conversion, SpatialSink Debug, cpal thread safety
- Gamelogic: player.rs crc mutability (15 fields), ai_player.rs Bool→bool, stealth_upgrade type fixes, pathfinder stubs, contain module xfer stubs, slowdeathbehavior type fixes
- control_bar.rs: added missing fn update_context_observer header (orphaned code at depth=1 inside impl)
- terrain_visual.rs: removed extra closing brace
- game_engine + gamelogic crates compile clean (0 errors)

+++ GeneralsRust/Code/GameEngine/GameLogic/src/scripting/engine.rs
@@ -1650,7 +1650,7 @@ impl ScriptEngine {
 
     fn template_matches_kind(template: &EngineThingTemplate, kind: KindOf) -> bool {
         for idx in kind_of_indices(kind) {
-            if template.is_kind_of(*idx) {
+            if template.is_kind_of((*idx) as u64) {
                 return true;
             }
         }
